requirements.txt
boto3
botocore
pydantic
PyJWT
bcrypt
python-jose
email-validator
